chore(x/upgrade): move tally_test to keeper_test
