+++ x/upgrade/keeper_test.go
@@ -10,9 +10,16 @@ import (
 	testutil "github.com/celestiaorg/celestia-app/v2/test/util"
 	"github.com/celestiaorg/celestia-app/v2/x/upgrade"
 	"github.com/celestiaorg/celestia-app/v2/x/upgrade/types"
+	"github.com/cosmos/cosmos-sdk/store"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/tendermint/tendermint/libs/log"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+	tmversion "github.com/tendermint/tendermint/proto/tendermint/version"
+	tmdb "github.com/tendermint/tm-db"
 )
 
 func TestGetVotingPowerThreshold(t *testing.T) {
@@ -76,4 +83,254 @@ func TestResetTally(t *testing.T) {
 	resp, err = upgradeKeeper.VersionTally(ctx, &types.QueryVersionTallyRequest{Version: 2})
 	require.NoError(t, err)
 	assert.Equal(t, uint64(0), resp.VotingPower)
+}
+
+func TestSignalVersion(t *testing.T) {
+	upgradeKeeper, ctx, _ := setup(t)
+	goCtx := sdk.WrapSDKContext(ctx)
+	t.Run("should return an error if the signal version is less than the current version", func(t *testing.T) {
+		_, err := upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+			ValidatorAddress: testutil.ValAddrs[0].String(),
+			Version:          0,
+		})
+		assert.Error(t, err)
+		assert.ErrorIs(t, err, types.ErrInvalidVersion)
+	})
+	t.Run("should return an error if the signal version is greater than the next version", func(t *testing.T) {
+		_, err := upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+			ValidatorAddress: testutil.ValAddrs[0].String(),
+			Version:          3,
+		})
+		assert.Error(t, err)
+		assert.ErrorIs(t, err, types.ErrInvalidVersion)
+	})
+	t.Run("should return an error if the validator was not found", func(t *testing.T) {
+		_, err := upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+			ValidatorAddress: testutil.ValAddrs[4].String(),
+			Version:          2,
+		})
+		require.Error(t, err)
+		require.ErrorIs(t, err, stakingtypes.ErrNoValidatorFound)
+	})
+	t.Run("should not return an error if the signal version and validator are valid", func(t *testing.T) {
+		_, err := upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+			ValidatorAddress: testutil.ValAddrs[0].String(),
+			Version:          2,
+		})
+		require.NoError(t, err)
+
+		res, err := upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+			Version: 2,
+		})
+		require.NoError(t, err)
+		require.EqualValues(t, 40, res.VotingPower)
+		require.EqualValues(t, 100, res.ThresholdPower)
+		require.EqualValues(t, 120, res.TotalVotingPower)
+	})
+}
+
+func TestTallyingLogic(t *testing.T) {
+	upgradeKeeper, ctx, mockStakingKeeper := setup(t)
+	goCtx := sdk.WrapSDKContext(ctx)
+	_, err := upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[0].String(),
+		Version:          0,
+	})
+	require.Error(t, err)
+	_, err = upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[0].String(),
+		Version:          3,
+	})
+	require.Error(t, err)
+
+	_, err = upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[0].String(),
+		Version:          2,
+	})
+	require.NoError(t, err)
+
+	res, err := upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+		Version: 2,
+	})
+	require.NoError(t, err)
+	require.EqualValues(t, 40, res.VotingPower)
+	require.EqualValues(t, 100, res.ThresholdPower)
+	require.EqualValues(t, 120, res.TotalVotingPower)
+
+	_, err = upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[2].String(),
+		Version:          2,
+	})
+	require.NoError(t, err)
+
+	res, err = upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+		Version: 2,
+	})
+	require.NoError(t, err)
+	require.EqualValues(t, 99, res.VotingPower)
+	require.EqualValues(t, 100, res.ThresholdPower)
+	require.EqualValues(t, 120, res.TotalVotingPower)
+
+	_, err = upgradeKeeper.TryUpgrade(goCtx, &types.MsgTryUpgrade{})
+	require.NoError(t, err)
+	shouldUpgrade, version := upgradeKeeper.ShouldUpgrade()
+	require.False(t, shouldUpgrade)
+	require.Equal(t, uint64(0), version)
+
+	// we now have 101/120
+	_, err = upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[1].String(),
+		Version:          2,
+	})
+	require.NoError(t, err)
+
+	_, err = upgradeKeeper.TryUpgrade(goCtx, &types.MsgTryUpgrade{})
+	require.NoError(t, err)
+	shouldUpgrade, version = upgradeKeeper.ShouldUpgrade()
+	require.True(t, shouldUpgrade)
+	require.Equal(t, uint64(2), version)
+	// update the version to 2
+	ctx = ctx.WithBlockHeader(tmproto.Header{
+		Version: tmversion.Consensus{
+			Block: 1,
+			App:   2,
+		},
+	})
+	goCtx = sdk.WrapSDKContext(ctx)
+
+	_, err = upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[0].String(),
+		Version:          3,
+	})
+	require.NoError(t, err)
+
+	res, err = upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+		Version: 2,
+	})
+	require.NoError(t, err)
+	require.EqualValues(t, 60, res.VotingPower)
+	require.EqualValues(t, 100, res.ThresholdPower)
+	require.EqualValues(t, 120, res.TotalVotingPower)
+
+	// remove one of the validators from the set
+	delete(mockStakingKeeper.validators, testutil.ValAddrs[1].String())
+	// the validator had 1 voting power, so we deduct it from the total
+	mockStakingKeeper.totalVotingPower = mockStakingKeeper.totalVotingPower.SubRaw(1)
+
+	res, err = upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+		Version: 2,
+	})
+	require.NoError(t, err)
+	require.EqualValues(t, 59, res.VotingPower)
+	require.EqualValues(t, 100, res.ThresholdPower)
+	require.EqualValues(t, 119, res.TotalVotingPower)
+
+	// That validator should not be able to signal a version
+	_, err = upgradeKeeper.SignalVersion(goCtx, &types.MsgSignalVersion{
+		ValidatorAddress: testutil.ValAddrs[1].String(),
+		Version:          2,
+	})
+	require.Error(t, err)
+
+	upgradeKeeper.ResetTally(ctx)
+	res, err = upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+		Version: 2,
+	})
+	require.NoError(t, err)
+	require.EqualValues(t, 0, res.VotingPower)
+}
+
+func TestEmptyStore(t *testing.T) {
+	upgradeKeeper, ctx, _ := setup(t)
+	goCtx := sdk.WrapSDKContext(ctx)
+
+	res, err := upgradeKeeper.VersionTally(goCtx, &types.QueryVersionTallyRequest{
+		Version: 2,
+	})
+	require.NoError(t, err)
+	require.EqualValues(t, 0, res.VotingPower)
+	// 120 is the summation in voting power of the four validators
+	require.EqualValues(t, 120, res.TotalVotingPower)
+}
+
+func TestThresholdVotingPower(t *testing.T) {
+	upgradeKeeper, ctx, mockStakingKeeper := setup(t)
+
+	for _, tc := range []struct {
+		total     int64
+		threshold int64
+	}{
+		{total: 1, threshold: 1},
+		{total: 2, threshold: 2},
+		{total: 3, threshold: 3},
+		{total: 6, threshold: 5},
+		{total: 59, threshold: 50},
+	} {
+		mockStakingKeeper.totalVotingPower = sdkmath.NewInt(tc.total)
+		threshold := upgradeKeeper.GetVotingPowerThreshold(ctx)
+		require.EqualValues(t, tc.threshold, threshold.Int64())
+	}
+}
+
+func setup(t *testing.T) (upgrade.Keeper, sdk.Context, *mockStakingKeeper) {
+	upgradeStore := sdk.NewKVStoreKey(types.StoreKey)
+	db := tmdb.NewMemDB()
+	stateStore := store.NewCommitMultiStore(db)
+	stateStore.MountStoreWithDB(upgradeStore, storetypes.StoreTypeIAVL, nil)
+	require.NoError(t, stateStore.LoadLatestVersion())
+	mockCtx := sdk.NewContext(stateStore, tmproto.Header{
+		Version: tmversion.Consensus{
+			Block: 1,
+			App:   1,
+		},
+	}, false, log.NewNopLogger())
+	mockStakingKeeper := newMockStakingKeeper(
+		map[string]int64{
+			testutil.ValAddrs[0].String(): 40,
+			testutil.ValAddrs[1].String(): 1,
+			testutil.ValAddrs[2].String(): 59,
+			testutil.ValAddrs[3].String(): 20,
+		},
+	)
+
+	upgradeKeeper := upgrade.NewKeeper(upgradeStore, mockStakingKeeper)
+	return upgradeKeeper, mockCtx, mockStakingKeeper
+}
+
+var _ upgrade.StakingKeeper = (*mockStakingKeeper)(nil)
+
+type mockStakingKeeper struct {
+	totalVotingPower sdkmath.Int
+	validators       map[string]int64
+}
+
+func newMockStakingKeeper(validators map[string]int64) *mockStakingKeeper {
+	totalVotingPower := sdkmath.NewInt(0)
+	for _, power := range validators {
+		totalVotingPower = totalVotingPower.AddRaw(power)
+	}
+	return &mockStakingKeeper{
+		totalVotingPower: totalVotingPower,
+		validators:       validators,
+	}
+}
+
+func (m *mockStakingKeeper) GetLastTotalPower(_ sdk.Context) sdkmath.Int {
+	return m.totalVotingPower
+}
+
+func (m *mockStakingKeeper) GetLastValidatorPower(_ sdk.Context, addr sdk.ValAddress) int64 {
+	addrStr := addr.String()
+	if power, ok := m.validators[addrStr]; ok {
+		return power
+	}
+	return 0
+}
+
+func (m *mockStakingKeeper) GetValidator(_ sdk.Context, addr sdk.ValAddress) (validator stakingtypes.Validator, found bool) {
+	addrStr := addr.String()
+	if _, ok := m.validators[addrStr]; ok {
+		return stakingtypes.Validator{Status: stakingtypes.Bonded}, true
+	}
+	return stakingtypes.Validator{}, false
 }
